Hardcode the serial port to be GPIO RS485.
While not appropriate for upstreaming in it's present state,
this patch turns the serial port into a gpio controlled RS485 port.

Tested with a variety of modbus devices.  Ideally, this patch would
use the proper "struct serial_rs485" structures in the kernel and
and properly obey ioctls to switch in and out of rs485 mode.

At present however, it is simply hardcoded, and the RS485 ioctls are
simply ignored.